Rather Large Update
Pulls changes from RE2R and includes all the nonsense I had done as well

+++ reframework/autorun/randomizer/DestroyObjects.lua
@@ -3,7 +3,7 @@ DestroyObjects.isInit = false
 DestroyObjects.lastRemoval = os.time()
 
 function DestroyObjects.Init()
-    if not DestroyObjects.isInit then
+    if Archipelago.IsConnected() and not DestroyObjects.isInit then
         DestroyObjects.isInit = true
         DestroyObjects.DestroyAll()
     end
